test(cli): keep skill checks compatible with Bun 1.0
